implemented react-alert
