+++ frontend/core/navbar/index.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import { signout } from 'modules/user/actions';
+import { signout } from 'modules/auth/actions';
 import Navbar from './navbar';
 
 const mapStateToProps = ({ user }) => ({
