+++ frontend/core/root_reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import user from 'modules/user/reducer';
+import user from 'modules/auth/reducer';
 import pool from 'modules/pool/reducer';
 import errors from 'core/errors/reducer';
