+++ package.json
@@ -31,6 +31,7 @@
     "node-sass": "^4.5.3",
     "prop-types": "^15.5.10",
     "react": "^15.4.2",
+    "react-alert": "^2.1.1",
     "react-autobind": "^1.0.6",
     "react-click-outside": "^2.3.1",
     "react-dom": "^15.4.2",
